# openapi.yaml
openapi: 3.0.3
info:
  title: Unikol API
  description: >
    API REST de la plateforme SaaS de gestion scolaire Unikol.
    Toutes les routes (sauf /auth/*) exigent un JWT Bearer valide.
    Le tenant (SchoolId) est dérivé du claim `schoolId` du token, jamais d'un paramètre de requête.
    Référence fonctionnelle complète : docs/Volume_4_API_Design.md
  version: "1.1.0"
servers:
  - url: https://api.sama-ecole.sn/v1
    description: Production
  - url: https://staging-api.sama-ecole.sn/v1
    description: Staging
  - url: http://localhost:5000/v1
    description: Développement local
security:
  - bearerAuth: []
...
paths:
...
  /teachers:
    get:
      tags: [Teachers]
      summary: Lister les enseignants (pagination, recherche)
      parameters:
        - name: page
          in: query
          schema: { type: integer, default: 1 }
        - name: pageSize
          in: query
          schema: { type: integer, default: 20 }
        - name: search
          in: query
          schema: { type: string }
      responses:
        '200':
          description: Liste paginée des enseignants
          content:
            application/json:
              schema: { $ref: '#/components/schemas/PaginatedTeachers' }
        '403': { description: Rôle non autorisé (Volume_7_Security.md « Enseignants ») }
    post:
      tags: [Teachers]
      summary: Créer un enseignant (matricule généré à l'enregistrement, ticket JGK-D03)
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/TeacherCreateRequest' }
      responses:
        '201':
          description: Enseignant créé
          content:
            application/json:
              schema: { $ref: '#/components/schemas/TeacherResult' }
        '403': { description: Réservé au Directeur et au Secrétariat }
        '422': { $ref: '#/components/responses/ValidationError' }
  /teachers/export/pdf:
    get:
      tags: [Teachers]
      summary: « LISTE DES ENSEIGNANTS » en PDF (A4 paysage, sans pagination)
      description: >
        Pendant de GET /students/export/pdf pour le corps professoral. Mêmes rôles que la
        consultation de la liste (Volume_7_Security.md « Enseignants » : Super Admin, Directeur,
        Secrétariat) — un export n'ouvre aucune donnée que le rôle ne puisse déjà lire à l'écran ;
        le rôle Finance en est donc exclu, contrairement à l'export des élèves. Opération journalisée
        dans les journaux d'audit (Volume_7_Security.md §7).
      parameters:
        - name: status
          in: query
          description: Filtre optionnel sur le statut. Omis = tous les statuts.
          schema:
            type: string
            enum: [Active, Suspended, Blocked]
      responses:
        '200':
          description: La liste des enseignants au format PDF
          content:
            application/pdf:
              schema: { type: string, format: binary }
        '403': { description: Rôle non autorisé (Volume_7_Security.md « Enseignants ») }
  /teachers/{id}:
    get:
      tags: [Teachers]
      summary: Fiche enseignant complète (ticket JGK-D04)
      description: >-
        Matières qualifiées, attributions classe/matière groupées par année scolaire — cette
        dernière liste EST l'historique, une attribution n'étant jamais réécrite d'une année sur l'autre.
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        '200':
          description: Fiche enseignant
          content:
            application/json:
              schema: { $ref: '#/components/schemas/TeacherProfile' }
        '403': { description: Rôle non autorisé (Volume_7_Security.md « Enseignants ») }
        '404': { description: Enseignant introuvable dans l'établissement courant }
    put:
      tags: [Teachers]
      summary: Corriger la fiche d'un enseignant (état civil, contact, matières qualifiées)
      description: >-
        Réservé au Directeur et au Secrétariat. Matricule et rattachement d'un compte de connexion
        (userId) non modifiables ici (voir TeacherUpdateRequest).
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/TeacherUpdateRequest' }
      responses:
        '200': { description: Fiche mise à jour, content: { application/json: { schema: { $ref: '#/components/schemas/Teacher' } } } }
        '403': { description: Réservé au Directeur et au Secrétariat }
        '404': { description: Enseignant introuvable dans l'établissement courant }
        '409': { description: Jeton de concurrence périmé }
        '422': { $ref: '#/components/responses/ValidationError' }
    delete:
      tags: [Teachers]
      summary: Archiver (soft delete) une fiche enseignant créée par pure erreur de saisie
      description: 'Refusée en 409 dès qu''une attribution classe/matière/année existe déjà.'
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
        - name: rowVersion
          in: query
          required: true
          schema: { type: integer, format: int64 }
      responses:
        '204': { description: Fiche archivée }
        '403': { description: Réservé au Directeur et au Secrétariat }
        '404': { description: Enseignant introuvable dans l'établissement courant }
        '409': { description: 'Jeton de concurrence périmé, ou une attribution existe déjà' }
  /teachers/{id}/assignments:
    post:
      tags: [Teachers]
      summary: Attribuer une classe/matière à l'enseignant sur l'année scolaire active (ticket JGK-D04)
      description: >-
        Aucune année scolaire dans le corps de la requête : l'année ACTIVE est résolue serveur, jamais
        choisie par le client (même règle que POST /enrollments).
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/TeacherAssignmentCreateRequest' }
      responses:
        '201':
          description: Attribution créée
          content:
            application/json:
              schema: { $ref: '#/components/schemas/TeacherAssignmentResult' }
        '403': { description: Réservé au Directeur et au Secrétariat }
        '404': { description: Enseignant introuvable dans l'établissement courant }
        '409': { description: Cette attribution (enseignant, classe, matière, année) existe déjà }
        '422': { $ref: '#/components/responses/ValidationError' }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
  responses:
...
    ValidationError:
      description: Erreur de validation
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
...
  schemas:
...
    ErrorResponse:
      type: object
      description: Format normalisé — voir docs/Volume_4_API_Design.md §0.4
      properties:
        code: { type: string, example: "VALIDATION_ERROR" }
        message: { type: string }
        details:
          type: array
          items:
            type: object
            properties:
              field: { type: string }
              reason: { type: string }
        traceId: { type: string }
...
    TeacherCreateRequest:
      type: object
      required: [fullName, email, subjectIds]
      properties:
        fullName: { type: string }
        email: { type: string, format: email }
        phone: { type: string }
        subjectIds: { type: array, items: { type: string, format: uuid } }
        userId:
          type: string
          format: uuid
          description: >-
            Compte de connexion (rôle Enseignant, même école) à rattacher à la fiche (ticket JGK-D06).
            Optionnel ; requis pour que cet enseignant puisse faire l'appel de ses classes.
    TeacherResult:
      type: object
      properties:
        id: { type: string, format: uuid }
        matricule: { type: string }
    Teacher:
      type: object
      properties:
        id: { type: string, format: uuid }
        matricule: { type: string }
        fullName: { type: string }
        email: { type: string, format: email }
        phone: { type: string }
        status: { type: string, enum: [Active, Suspended, Blocked] }
        subjects: { type: array, items: { type: string } }
        rowVersion: { type: integer, format: int64, description: Jeton xmin — requis pour PUT/DELETE (verrou optimiste) }
    TeacherUpdateRequest:
      type: object
      required: [fullName, email, subjectIds, rowVersion]
      description: >-
        Le matricule et le rattachement d'un compte de connexion (userId) ne sont jamais modifiables
        ici — absents de ce contrat (voir TeacherCreateRequest pour la création initiale).
      properties:
        fullName: { type: string, maxLength: 200 }
        email: { type: string, format: email, maxLength: 255 }
        phone: { type: string, maxLength: 30 }
        birthPlace: { type: string, maxLength: 200 }
        photoUrl: { type: string, format: uri }
        subjectIds: { type: array, items: { type: string, format: uuid } }
        rowVersion: { type: integer, format: int64 }
    PaginatedTeachers:
      type: object
      properties:
        items: { type: array, items: { $ref: '#/components/schemas/Teacher' } }
        totalCount: { type: integer }
        page: { type: integer }
        pageSize: { type: integer }
    TeacherAssignment:
      type: object
      properties:
        id: { type: string, format: uuid }
        classroomName: { type: string }
        subjectName: { type: string }
        schoolYearId: { type: string, format: uuid }
        schoolYearLabel: { type: string }
        isActiveSchoolYear: { type: boolean }
    TeacherProfile:
      allOf:
        - $ref: '#/components/schemas/Teacher'
        - type: object
          properties:
            assignments: { type: array, items: { $ref: '#/components/schemas/TeacherAssignment' } }
    TeacherAssignmentCreateRequest:
      type: object
      required: [classroomId, subjectId]
      properties:
        classroomId: { type: string, format: uuid }
        subjectId: { type: string, format: uuid }
    TeacherAssignmentResult:
      type: object
      properties:
        id: { type: string, format: uuid }
        classroomId: { type: string, format: uuid }
        subjectId: { type: string, format: uuid }
        schoolYearId: { type: string, format: uuid }
...